+++ packages/Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-version = "0.9.20"
+version = "0.9.21"
 resolver = "2"
 members = [
     "sidan-csl-rs",
