+++ packages/sidan-csl-rs/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "sidan-csl-rs"
-version = "0.9.20"
+version = "0.9.21"
 edition = "2021"
 license = "Apache-2.0"
 description = "Wrapper around the cardano-serialization-lib for easier transaction building, heavily inspired by cardano-cli APIs"
