+++ packages/whisky-examples/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "whisky-examples"
-version = "0.9.20"
+version = "0.9.21"
 edition = "2021"
 license = "Apache-2.0"
 description = "The Cardano Rust SDK, inspired by MeshJS"
@@ -13,4 +13,4 @@ path = "src/server.rs"
 actix-cors = "0.7.0"
 actix-web = "4.9.0"
 serde = "1.0.209"
-whisky = { version = "=0.9.20", path = "../whisky" }
+whisky = { version = "=0.9.21", path = "../whisky" }
