+++ packages/whisky/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "whisky"
-version = "0.9.20"
+version = "0.9.21"
 edition = "2021"
 license = "Apache-2.0"
 description = "The Cardano Rust SDK, inspired by MeshJS"
@@ -24,7 +24,7 @@ pallas-codec = { version = "0.30.2", features = ["num-bigint"] }
 pallas-primitives = "0.31.0"
 pallas-traverse = "0.31.0"
 maestro-rust-sdk = "1.1.3"
-sidan-csl-rs = { version = "0.9.20", path = "../sidan-csl-rs" }
+sidan-csl-rs = { version = "0.9.21", path = "../sidan-csl-rs" }
 reqwest = "0.12.5"
 tokio = { version = "1.38.0", features = ["macros", "rt-multi-thread"] }
